+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- ActiveStorage integration.
+
 ## 0.1.0 - 2025-08-10
 
 - Initial release extracted from Parklife 0.7.0.